feat: 发布版本1.0.2 - 完善采购管理模块
- 修复PurReceiptServiceImpl和PurOrderService中的类型不匹配问题
- 添加采购对账功能模块
- 完善仓库管理功能
- 优化供应商商品管理
- 完善采购退货功能
- 添加地区和库位管理
- 修复数据库字段类型一致性问题
- 优化前端组件和路由配置

+++ src/main/java/com/example/vliascrm/dto/PurReturnItemDto.java
@@ -80,7 +80,7 @@ public class PurReturnItemDto {
     /**
      * 退货数量
      */
-    private Integer quantity;
+    private BigDecimal quantity;
 
     /**
      * 总金额
@@ -127,11 +127,11 @@ public class PurReturnItemDto {
     /**
      * 可退数量（基于入库数量）
      */
-    private Integer availableQuantity;
+    private BigDecimal availableQuantity;
 
     /**
      * 已退数量
      */
-    private Integer returnedQuantity;
+    private BigDecimal returnedQuantity;
 }
